make service transactional

+++ src/main/java/com/example/todoservice/todo/TodoService.java
@@ -1,11 +1,10 @@
 package com.example.todoservice.todo;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class TodoService {
@@ -15,39 +14,40 @@ public class TodoService {
         this.todoRepository = todoRepository;
     }
 
+    @Transactional(readOnly = true)
     public List<Todo> findByUsername(String username) {
-        List<Todo> todos = new ArrayList<>();
-        todoRepository.findByUsername(username).forEach(todo -> todos.add(todo));
-        return todos;
+        return todoRepository.findByUsername(username);
     }
 
+    @Transactional(readOnly = true)
     public Optional<Todo> findById(Long id) {
         return todoRepository.findById(id);
     }
 
-    public Todo findByIdAndUsername(Long id, String username) {
-        return todoRepository.findByIdAndUsername(id, username);
-    }
-
+    @Transactional
     public boolean deleteTodoById(Long id) {
-        try {
+        Optional<Todo> todoToDelete = todoRepository.findById(id);
+
+        if (todoToDelete.isPresent()) {
             todoRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
+            return true;
+        } else {
             return false;
         }
-        return true;
     }
 
+    @Transactional
     public Todo save(String username, Todo todo) {
         return save(Long.valueOf(-1), username, todo);
     }
 
+    @Transactional
     public Todo save(Long id, String username, Todo todo) {
         if (id.equals(Long.valueOf(-1))) {
             todo.setUsername(username);
             return todoRepository.save(todo);
         } else {
-            Todo todoToUpdate = findByIdAndUsername(id, username);
+            Todo todoToUpdate = todoRepository.findByIdAndUsername(id, username);
             // TODO: use mapper
             todoToUpdate.setDescription(todo.getDescription());
             todoToUpdate.setTargetDate(todo.getTargetDate());
